Develop

See merge request s08-bigdata-recom-sub2/S08P22C203!25

+++ backend/src/main/java/com/ssafy/ododok/db/model/Board.java
@@ -19,7 +19,7 @@ public class Board {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int boardId;
+    private Long boardId;
 
     @Column(nullable = false)
     private String boardTitle;
